requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
